Enforces valid UTF8 keys on DynamoDB back end
The use of non-UTF8 keys with the DynamoDB back-end causes a failure
deep within the AWS request (de)serialising code, presenting a
non-obvious failure to the user.

This change adds validation at the DynamoDB `Backend` interface level
to require that all keys are valid UTF8 strings. It also adds a warning
to the `Backend` interface declaration that the keys may be constrained
to valid UTF-8.

Other changes include:
 * Updating the `Backend` conformance test suite to not present binary
   keys to the various backend implementations.
 * Adding a `region` value to the DynamoDB configuration test input
 * Adding missing imports to `_test` files.
 * Updating build instructions in README